THIS IS BROKEN DO NOT TEST
This fixes the GUI on the gem refiner.  The client is updated for progress and the gui actually has a work cycle not just a time check before processing the job.  This gives us time to show a proper progress bar on the GUI and it works when you walk away from it.

This commit also removes ttcore by request.  I copied the class we used into our JSON tool (with permission I am told), but we have a new bug.  For some reason the oredict lookup is not working correctly on loading itemCoal and I am too tired now to deal with it.

I will be putting the other GUI's to work and getting them to cycle like proper tile's tomorrow, this includes the soil.

@jaredlll08 look at the JSON bit, it is copy for copy with what was in the original mod.